requirements.txt
yattag
pyephem
fabric
requests
python-dateutil
pillow
opencv-python
fits2image
boto3
